fix 268: dataloader workers leak semaphore objects
